requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
bcrypt>=4.0
PyJWT>=2.0
